audioDevice moved out to its own capability
