+++ packages/teams-js/src/public/audioVisualDevice.ts
@@ -30,7 +30,7 @@ import {
   generateGUID,
   InputFunction,
 } from '../internal/utils';
-import { FrameContexts, HostClientType } from './constants';
+import { FrameContexts, HostClientType, MediaType } from './constants';
 import { ErrorCode, SdkError } from './interfaces';
 import { media } from './media';
 import { runtime } from './runtime';
@@ -197,16 +197,6 @@ export namespace audioVisualDevice {
      * chunk sequence number
      */
     chunkSequence: number;
-  }
-
-  /**
-   * Specifies the type of Media
-   */
-  export enum MediaType {
-    Image = 1,
-    Video = 2,
-    VideoAndImage = 3,
-    Audio = 4,
   }
 
   // This should not trigger the "refresh the app scenario" because this is for setting things up
@@ -685,75 +675,6 @@ export namespace audioVisualDevice {
       export function isSupported(): boolean {
         return runtime.supports.media.camera.barcode ? true : false;
       }
-    }
-  }
-
-  export namespace audio {
-    /**
-     * Input parameter supplied to the select Media API
-     */
-    export interface AudioInputs {
-      /**
-       * Only one media type can be selected at a time
-       */
-      mediaType: MediaType.Audio;
-
-      /**
-       * max limit of media allowed to be selected in one go, current max limit is 10 set by office lens.
-       */
-      maxMediaCount: number;
-
-      /**
-       * Additional properties for audio capture flows.
-       */
-      audioProps?: AudioProps;
-    }
-
-    /**
-     *  All properties in AudioProps are optional and have default values in the platform
-     */
-    export interface AudioProps {
-      /**
-       * Optional; the maximum duration in minutes after which the recording should terminate automatically
-       * Default value is defined by the platform serving the API.
-       */
-      maxDuration?: number;
-    }
-
-    export function selectAudio(audioInput: AudioInputs): Promise<media.Media[]> {
-      ensureInitialized(FrameContexts.content, FrameContexts.task);
-
-      // Probably should clean this up, no reason to use this structure anymore
-      const wrappedFunction: InputFunction<media.Media[]> = () =>
-        new Promise<[SdkError, media.Media[]]>(resolve => {
-          if (!isCurrentSDKVersionAtLeast(mediaAPISupportVersion)) {
-            throw { errorCode: ErrorCode.OLD_PLATFORM };
-          }
-
-          if (!validateSelectMediaInputs(audioInput)) {
-            throw { errorCode: ErrorCode.INVALID_ARGUMENTS };
-          }
-
-          const params = [audioInput];
-          // What comes back from native at attachments would just be objects and will be missing getMedia method on them.
-          resolve(sendMessageToParentAsync<[SdkError, media.Media[]]>('selectMedia', params));
-        }).then(([err, localAttachments]: [SdkError, media.Media[]]) => {
-          // Media Attachments are final response to selectMedia
-          if (!localAttachments) {
-            throw err;
-          }
-          const mediaArray: media.Media[] = [];
-          for (const attachment of localAttachments) {
-            mediaArray.push(new media.Media(attachment));
-          }
-          return mediaArray;
-        });
-
-      return callCallbackWithErrorOrResultFromPromiseAndReturnPromise<media.Media[]>(wrappedFunction);
-    }
-
-    export function isSupported(): boolean {
-      return runtime.supports.media.audio ? true : false;
     }
   }
 }
